requirements.txt
streamlit
web3
pandas
requests
python-dotenv
altair